+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 
 - CHANGED: Rescue all SystemCallError instead of a few Errno errors (GH-212). [Thanks @mat813]
 
+- CHANGED: Removed deprecated method Whois::Client#query.
+
 - FIXED: whois.ua should support status `ok`.
 
 - FIXED: whois.registry.net.za crashes when the registrant doesn't have a name.